+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
+## Unreleased
+
+### Added 
+
+* Added support for the PostgreSQL [`Point` type][pg-point].
+
+[pg-point]: https://www.postgresql.org/docs/9.3/static/datatype-geometric.html
+
 ## [1.2.0] - 2018-04-06
 
 ### Added